role and permission

+++ resources/views/product/index.blade.php
@@ -1,10 +1,21 @@
 @extends('layouts.frontend')
 @section('title', 'รายการสินค้าทั้งหมด')
 @section('content')
+    <div class="row">
+        <div class="col-lg-12 margin-tb">
+            <div class="pull-left">
+                <h2>สินค้า</h2>
+            </div>
+            <div class="pull-right">
+                @can('product-create')
+                    <a class="btn btn-success" href="{{ route('products.create') }}"> เพิ่มสินค้าใหม่ </a>
+                @endcan
+            </div>
+        </div>
+    </div>
     <section class="py-5 container">
         <div class="row">
             <div class="col-lg-12  mx-auto">
-
                 <table class="table table-hover">
                     <thead>
                         <tr>
@@ -24,25 +35,23 @@
                                 <td scope="col">{{ $product->name }}</td>
                                 <td scope="col" style="width: 50%">{{ $product->price }}</td>
                                 <td scope="col">
-                                    <a href="{{ route('products.edit', $product->id) }}"
-                                        class="btn btn-info btn-sm">แก้ไข</a>
+                                    @can('product-edit')
+                                        <a class="btn btn-primary" href="{{ route('products.edit', $product->id) }}">แก้ไข</a>
+                                    @endcan
 
-                                    {{-- <a href="{{ route('products.destroy', $product->id) }}" class="btn btn-danger btn-sm"
-                                        onclick="return confirm('ยืนยันการลบหรือไม่')">ลบ
-                                    </a> --}}
                                 </td>
                                 <td scope="col">
-                                    {!! Form::open([
-                                        'route' => ['products.destroy', $product->id],
-                                        'method' => 'DELETE',
-                                    ]) !!}
-                                    {!! Form::submit('ลบข้อมูล', [
-                                        'class' => 'btn btn-danger btn-sm',
-                                        'onclick' => "return confirm('ยืนยันการลบหรือไม่')",
-                                    ]) !!}
-
-                                    {!! Form::close() !!}
-
+                                    @can('product-delete')
+                                        {!! Form::open([
+                                            'route' => ['products.destroy', $product->id],
+                                            'method' => 'DELETE',
+                                        ]) !!}
+                                        {!! Form::submit('ลบข้อมูล', [
+                                            'class' => 'btn btn-danger btn-sm',
+                                            'onclick' => "return confirm('ยืนยันการลบหรือไม่')",
+                                        ]) !!}
+                                        {!! Form::close() !!}
+                                    @endcan
                                 </td>
                             </tr>
                         </tbody>
